Passing 6502_functional_test! ADC/SBC overhaul 2
This commit again reworks addition logic to correctly perform packed BCD the way the 6502 does it. The solution was built from scratch because most guides I found were not correct (at least as far as the test ROM was concerned). With this we now pass 6502_functional_test by Klaus Dormann.